Asteroids must be mined now
Player has to keep laser trained on an asteroid for 1s to
mine its material. Asteroids are now uncolored until they
are mined, and do not give resources to the star if they
crash into it before being mined.

Also fixed bug with the star incorrect time penalty so it
counts down, not up when you get the wrong thing.